remove unnecessary directories within src/pages
