+++ src/pages/player-profile.tsx
