+++ src/pages/username-select.tsx
@@ -1,8 +1,7 @@
-import { Stack, type SxProps } from "@mui/material";
+import { Stack } from "@mui/material";
 import { type FC, useEffect } from "react";
 import { useRouter } from "next/router.js";
 import { signIn, useSession } from "next-auth/react";
-
 import { AddUsernameTable } from "~/components/tables/index.js";
 import { RowCreateUsername } from "~/components/tables/components/index.js";
 import { isVerifiedSession } from "~/utils/session.js";
@@ -10,21 +9,7 @@ import { pages } from "~/router.js";
 import { api } from "~/utils/api.js";
 import { useBBWallet } from "~/service/bb-wallet.js";
 import Page from "~/components/page/page.jsx";
-
-// TODO: move styles to another file
-const styles = {
-  wrapper: {
-    // TODO: add color to theme
-    backgroundColor: "#9E9E9E1E",
-    width: "100%",
-    alignItems: "center",
-  },
-  container: {
-    position: "relative",
-    overflow: "hidden",
-    height: "100vh",
-  } satisfies SxProps,
-};
+import { styles } from '../styles/pages/username-select';
 
 const UsernameSelect: FC = () => {
   const router = useRouter();
