+++ src/styles/pages/username-select.ts
@@ -1,0 +1,15 @@
+import { type SxProps } from "@mui/material";
+
+export const styles = {
+    wrapper: {
+      // TODO: add color to theme
+      backgroundColor: "#9E9E9E1E",
+      width: "100%",
+      alignItems: "center",
+    },
+    container: {
+      position: "relative",
+      overflow: "hidden",
+      height: "100vh",
+    } satisfies SxProps,
+  };
